feat(sql-client,main-node): postgres adapter + DATABASE_URL backend selector
Adds a postgres.js-backed SqlClient implementation alongside the existing
better-sqlite3 / D1 adapters. main-node now picks the backend at startup
from DATABASE_URL (postgres:// → PG) or DATABASE_PATH (→ better-sqlite3).
Same code path everywhere — every store package + the event log work
identically on both backends because they all ride on SqlClient.

Why two databases off one port:
  - SQLite is the simplest local-dev story; one file, no daemon.
  - Postgres unlocks multi-process and real durability for prod CFless.
  - Both are valid Fly/Render/anywhere targets; operator picks at deploy.
  - Code stays a single port — no per-backend repo forks.

Implementation:

  packages/sql-client/src/adapters/postgres.ts — PostgresSqlClient.
    - Translates SQLite-style `?` placeholders to PG `$1, $2, ...`. Has
      a small scanner that preserves `?` inside SQL string literals
      ('...' / "...") and -- comments.
    - Maps client.batch to sql.begin (transaction). Same atomicity
      semantics as D1 batch + better-sqlite3 db.transaction.
    - Lazy-imports `postgres` (peer dep, optional). Package compiles +
      runs without postgres installed unless you actually create a
      PostgresSqlClient.
    - BIGINT type-parser registered at construction so postgres.js
      returns ms timestamps + counters as JS numbers (not strings).
      Repos that do new Date(row.created_at) just work without per-
      repo Number() casts.

  packages/sql-client/src/index.ts — top-level re-exports
    createPostgresSqlClient (driver-lazy, safe to expose).

  packages/sql-client/package.json — postgres ^3.4 added as optional
    peer dep.

  apps/main-node/src/index.ts:
    - DATABASE_URL takes precedence over DATABASE_PATH; postgres:// or
      postgresql:// → PG, otherwise sqlite.
    - Bootstrap DDL converted INTEGER → BIGINT for timestamp / seq /
      size_bytes columns (PG INTEGER is 32-bit, would overflow ms
      timestamps in 2038). SQLite ignores the type name distinction
      (BIGINT has INTEGER affinity).

  packages/event-log/src/sql/index.ts — same INTEGER → BIGINT for
    session_events.{seq,ts} and session_streams.{started_at,completed_at}.

  docs/cfless.md + .env.cfless.example — document DATABASE_URL.

Verified e2e on PG (docker postgres:16-alpine):
  1. main-node boots, /health reports backends.agents=postgres
  2. \dt in psql shows all 11 tables created
  3. POST /v1/vaults + /credentials (BIGINT timestamp roundtrip ✓)
  4. POST /v1/memory_stores + /memories
  5. POST /v1/agents + /sessions + /memory_stores binding
  6. user.message → agent reads /mnt/memory/prefs/style.md (returns
     seeded content) → writes /mnt/memory/prefs/from-pg.md
  7. chokidar watcher fires upsertFromEvent → PG `memories` row
     created → /v1/memory_stores/:id/memories returns both files
  8. SELECT * FROM memories in psql confirms the rows

Verification:
  pnpm typecheck:node  → all 3 (sandbox + main-node + oma-vault) pass
  pnpm test            → 191 failed (matches baseline, 0 delta)

+++ packages/event-log/src/sql/index.ts
@@ -252,10 +252,10 @@ export async function ensureSchema(sql: SqlClient): Promise<void> {
   await sql.exec(`
     CREATE TABLE IF NOT EXISTS session_events (
       session_id TEXT NOT NULL,
-      seq INTEGER NOT NULL,
+      seq BIGINT NOT NULL,
       type TEXT NOT NULL,
       data TEXT NOT NULL,
-      ts INTEGER NOT NULL,
+      ts BIGINT NOT NULL,
       PRIMARY KEY (session_id, seq)
     );
   `);
@@ -269,8 +269,8 @@ export async function ensureSchema(sql: SqlClient): Promise<void> {
       message_id TEXT NOT NULL,
       status TEXT NOT NULL,
       chunks_json TEXT NOT NULL DEFAULT '[]',
-      started_at INTEGER NOT NULL,
-      completed_at INTEGER,
+      started_at BIGINT NOT NULL,
+      completed_at BIGINT,
       error_text TEXT,
       PRIMARY KEY (session_id, message_id)
     );
